requirements.txt
Django==1.8.4
PyJWT==1.4.0
PyYAML==3.11
Pygments==2.0.2
argparse==1.2.1
coverage==3.7.1
dj-database-url==0.3.0
dj-static==0.0.6
django-excel==0.0.3
django-model-utils==2.4
django-push-notifications==1.4.1
django-rest-swagger==0.3.5
django-social-auth==0.7.28
django-toolbelt==0.0.1
djangorestframework==3.2.4
djangorestframework-jwt==1.7.2
future==0.15.2
gunicorn==19.3.0
httpie==0.9.3
httplib2==0.9.2
jdcal==1.2
nose==1.3.7
nose-cov==1.6
oauth2==1.9.0.post1
oauthlib==1.0.3
odfpy==0.9.6
openpyxl==2.2.2
psycopg2==2.6.1
pyexcel==0.1.7
pyexcel-io==0.0.9
pyexcel-ods==0.0.12
pyexcel-ods3==0.0.12
pyexcel-webio==0.0.4
pyexcel-xls==0.0.8
pyexcel-xlsx==0.0.9
python-coveralls==2.6.0
python-http-client==1.2.3
python-openid==2.2.5
python-social-auth==0.2.13
rednose==0.4.3
requests==2.7.0
requests-oauthlib==0.5.0
sendgrid==2.2.1
six==1.9.0
smtpapi==0.3.1
static3==0.6.1
texttable==0.8.4
wsgiref==0.1.2
xlrd==0.9.4
xlwt-future==0.8.0
